Replace a common interview form template with a component
InterviewInputsFactory is not using a generated factory/interface because I couldn't get it working with the name service that should be passed to InterviewInputs, was getting "Service '' is not defined." because of some mapping applied strangely when only one create() parameter was present probably.
