+++ site/app/Admin/Presenters/PresenterTemplates/Interviews/InterviewsInterview.php
@@ -3,19 +3,14 @@ declare(strict_types = 1);
 
 namespace MichalSpacekCz\Admin\Presenters\PresenterTemplates\Interviews;
 
+use MichalSpacekCz\Interviews\Interview;
 use Nette\Bridges\ApplicationLatte\Template;
-use Nette\Database\Row;
 
 class InterviewsInterview extends Template
 {
 
 	public string $pageTitle;
 
-	/** @var Row<mixed> */
-	public Row $interview;
-
-	public int $videoThumbnailWidth;
-
-	public int $videoThumbnailHeight;
+	public Interview $interview;
 
 }
