+++ site/app/Admin/Presenters/templates/Talks/common/talkForm.latte
@@ -36,7 +36,7 @@
 		<tr>
 			<th>{label videoHref /}</th><td>{input videoHref}</td>
 		</tr>
-		{include "formVideoThumbnails.latte", video: $talk?->getVideo(), videoThumbnailWidth: $videoThumbnailWidth, videoThumbnailHeight: $videoThumbnailHeight}
+		{include "../../../../../Media/Thumbnails/videoThumbnailsInputs.latte", video: $talk?->getVideo(), videoThumbnailWidth: $videoThumbnailWidth, videoThumbnailHeight: $videoThumbnailHeight}
 		<tr class="lighter">
 			<th class="align-top">{label videoEmbed /}</th><td>{input videoEmbed}<br><small><em>deprecated</em></small></td>
 		</tr>
