+++ site/app/Media/Thumbnails/videoThumbnailsInputs.latte
